+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/actions/FlexibleSearchToggleInEditorResultsAction.kt
@@ -26,8 +26,8 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.ToggleAction
 
 class FlexibleSearchToggleInEditorResultsAction : ToggleAction(
-    message("hybris.actions.in_editor_results"),
-    message("hybris.actions.in_editor_results.description"),
+    message("hybris.actions.in_editor.results"),
+    message("hybris.actions.in_editor.results.description"),
     HybrisIcons.Actions.TOGGLE_IN_EDITOR_RESULTS
 ) {