+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExExecuteAction.kt
@@ -21,6 +21,7 @@ import com.intellij.idea.plugin.hybris.actions.ExecuteStatementAction
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.impex.ImpexLanguage
+import com.intellij.idea.plugin.hybris.impex.editor.ImpExSplitEditor
 import com.intellij.idea.plugin.hybris.impex.editor.impexSplitEditor
 import com.intellij.idea.plugin.hybris.tools.remote.console.impl.HybrisImpexConsole
 import com.intellij.idea.plugin.hybris.tools.remote.execution.impex.ImpExExecutionClient
@@ -29,36 +30,24 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.application.readAction
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.Key
-import com.intellij.ui.AnimatedIcon
 import kotlinx.coroutines.launch
 
-class ImpExExecuteAction : ExecuteStatementAction<HybrisImpexConsole>(
+class ImpExExecuteAction : ExecuteStatementAction<HybrisImpexConsole, ImpExSplitEditor>(
     ImpexLanguage,
     HybrisImpexConsole::class,
     message("hybris.impex.actions.execute_query"),
     message("hybris.impex.actions.execute_query.description"),
     HybrisIcons.Console.Actions.EXECUTE
 ) {
-    override fun update(e: AnActionEvent) {
-        super.update(e)
 
-        val queryExecuting = e.impexSplitEditor()
-            ?.getUserData(KEY_QUERY_EXECUTING)
-            ?: false
-
-        e.presentation.isEnabledAndVisible = e.presentation.isEnabledAndVisible
-        e.presentation.isEnabled = e.presentation.isEnabledAndVisible && !queryExecuting
-        e.presentation.disabledIcon = if (queryExecuting) AnimatedIcon.Default.INSTANCE
-        else HybrisIcons.Console.Actions.EXECUTE
-    }
+    override fun fileEditor(e: AnActionEvent): ImpExSplitEditor? = e.impexSplitEditor()
 
     override fun processContent(e: AnActionEvent, content: String, editor: Editor, project: Project): String = e.impexSplitEditor()
         ?.virtualText
         ?: content
 
     override fun actionPerformed(e: AnActionEvent, project: Project, content: String) {
-        val fileEditor = e.impexSplitEditor()
+        val fileEditor = fileEditor(e)
         val context = ImpExExecutionContext(
             content = content
         )
@@ -82,9 +71,5 @@ class ImpExExecuteAction : ExecuteStatementAction<HybrisImpexConsole>(
                 console.print(result)
             }
         }
-    }
-
-    companion object {
-        private val KEY_QUERY_EXECUTING = Key.create<Boolean>("impex.query.execution.state")
     }
 }