+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/actions/PolyglotQueryExecuteAction.kt
@@ -36,36 +36,24 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.application.readAction
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.Key
 import com.intellij.psi.util.PsiTreeUtil
-import com.intellij.ui.AnimatedIcon
 import com.intellij.util.asSafely
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import org.apache.http.HttpStatus
 
-class PolyglotQueryExecuteAction : ExecuteStatementAction<HybrisPolyglotQueryConsole>(
+class PolyglotQueryExecuteAction : ExecuteStatementAction<HybrisPolyglotQueryConsole, PolyglotQuerySplitEditor>(
     PolyglotQueryLanguage,
     HybrisPolyglotQueryConsole::class,
     message("hybris.pgq.actions.execute_query"),
     message("hybris.pgq.actions.execute_query.description"),
     HybrisIcons.Console.Actions.EXECUTE
 ) {
-    override fun update(e: AnActionEvent) {
-        super.update(e)
 
-        val queryExecuting = e.polyglotQuerySplitEditor()
-            ?.getUserData(KEY_QUERY_EXECUTING)
-            ?: false
-
-        e.presentation.isEnabledAndVisible = e.presentation.isEnabledAndVisible
-        e.presentation.isEnabled = e.presentation.isEnabledAndVisible && !queryExecuting
-        e.presentation.disabledIcon = if (queryExecuting) AnimatedIcon.Default.INSTANCE
-        else HybrisIcons.Console.Actions.EXECUTE
-    }
+    override fun fileEditor(e: AnActionEvent): PolyglotQuerySplitEditor? = e.polyglotQuerySplitEditor()
 
     override fun actionPerformed(e: AnActionEvent, project: Project, content: String) {
-        val fileEditor = e.polyglotQuerySplitEditor() ?: return
+        val fileEditor = fileEditor(e) ?: return
         val itemType = e.getData(CommonDataKeys.PSI_FILE)
             ?.asSafely<PolyglotQueryFile>()
             ?.let { PsiTreeUtil.findChildOfType(it, PolyglotQueryTypeKeyName::class.java) }
@@ -225,7 +213,4 @@ class PolyglotQueryExecuteAction : ExecuteStatementAction<HybrisPolyglotQueryCon
             exec.invoke(coroutineScope, result)
         }
 
-    companion object {
-        private val KEY_QUERY_EXECUTING = Key.create<Boolean>("pgq.query.execution.state")
-    }
 }